requirements.txt
google-genai>=0.1.0
gTTS>=2.3.0
moviepy>=1.0.3
numpy>=1.20.0
pillow>=8.0.0
imageio>=2.9.0
imageio-ffmpeg>=0.4.0